upcoming: [DI-20933] - Renamed type name
