+++ packages/manager/src/features/CloudPulse/Dashboard/CloudPulseDashboard.tsx
@@ -15,9 +15,9 @@ import { RenderWidgets } from '../Widget/CloudPulseWidgetRenderer';
 
 import type { CloudPulseMetricsAdditionalFilters } from '../Widget/CloudPulseWidget';
 import type {
+  DateTimeWithPreset,
   JWETokenPayLoad,
   TimeDuration,
-  TimeDurationDate,
 } from '@linode/api-v4';
 
 export interface DashboardProperties {
@@ -34,7 +34,7 @@ export interface DashboardProperties {
   /**
    * time duration to fetch the metrics data in this widget
    */
-  duration: TimeDurationDate;
+  duration: DateTimeWithPreset;
 
   /**
    * optional timestamp to pass as react query param to forcefully re-fetch data
