+++ packages/manager/src/features/CloudPulse/Utils/ReusableDashboardFilterUtils.ts
@@ -1,10 +1,10 @@
-import { defaultTimeDuration } from '../shared/CloudPulseDateTimeRangePicker';
+import { defaultTimeDuration } from './CloudPulseDateTimePickerUtils';
 import { FILTER_CONFIG } from './FilterConfig';
 
 import type { DashboardProperties } from '../Dashboard/CloudPulseDashboard';
 import type { FilterValueType } from '../Dashboard/CloudPulseDashboardLanding';
 import type { CloudPulseMetricsAdditionalFilters } from '../Widget/CloudPulseWidget';
-import type { Dashboard, TimeDurationDate } from '@linode/api-v4';
+import type { Dashboard, DateTimeWithPreset } from '@linode/api-v4';
 
 /**
  * This interface is used to get method parameters for this utility
@@ -25,7 +25,7 @@ interface ReusableDashboardFilterUtilProps {
   /**
    * The selected time duration
    */
-  timeDuration?: TimeDurationDate;
+  timeDuration?: DateTimeWithPreset;
 }
 
 /**
