+++ packages/manager/src/features/CloudPulse/Widget/CloudPulseWidget.tsx
@@ -23,7 +23,7 @@ import { ZoomIcon } from './components/Zoomer';
 
 import type { FilterValueType } from '../Dashboard/CloudPulseDashboardLanding';
 import type { CloudPulseResources } from '../shared/CloudPulseResourcesSelect';
-import type { TimeDurationDate, Widgets } from '@linode/api-v4';
+import type { DateTimeWithPreset, Widgets } from '@linode/api-v4';
 import type {
   MetricDefinition,
   TimeDuration,
@@ -61,7 +61,7 @@ export interface CloudPulseWidgetProperties {
   /**
    * time duration to fetch the metrics data in this widget
    */
-  duration: TimeDurationDate;
+  duration: DateTimeWithPreset;
 
   /**
    * entity ids selected by user to show metrics for
